Wrap svgbob install in conditional

+++ src/current/netlify/build.sh
@@ -19,9 +19,10 @@ function build {
 
 # Handle Rust dependencies
 rustup default stable
-cargo install svgbob_cli
-ln -s /opt/buildhome/.cargo/bin/svgbob_cli /opt/buildhome/.cargo/bin/svgbob
-
+if [ ! -f /opt/buildhome/.cargo/bin/svgbob_cli ]; then
+	cargo install svgbob_cli
+	ln -s /opt/buildhome/.cargo/bin/svgbob_cli /opt/buildhome/.cargo/bin/svgbob
+fi;
 gem install bundler --silent
 bundle install --quiet
 build _config_cockroachdb.yml,_config_url.yml
